+++ src/angular-generate.ts
@@ -1,12 +1,12 @@
 import { Project } from './project';
 import * as vscode from 'vscode';
-import { getRunOutput, getStringFrom, openUri } from './utilities';
+import { getRunOutput, getStringFrom, openUri, replaceAll } from './utilities';
 import { getOutputChannel, writeError, writeIonic } from './extension';
 import { join } from 'path';
 import { existsSync } from 'fs';
 
 export async function angularGenerate(project: Project, angularType: string): Promise<void> {
-  const name = await vscode.window.showInputBox({
+  let name = await vscode.window.showInputBox({
     title: `New Angular ${angularType}`,
     placeHolder: `Enter name for new ${angularType}`,
   });
@@ -14,9 +14,10 @@ export async function angularGenerate(project: Project, angularType: string): Pr
   if (!name || name.length < 1) return;
 
   // CREATE src/app/test2/test2.component.ts
-  writeIonic(`Creating Angular ${angularType} named ${name}..`);
   try {
-    const out = await getRunOutput(`npx ionic generate ${angularType} '${name}'`, project.projectFolder());
+    name = replaceAll(name, ' ', '-').trim();
+    writeIonic(`Creating Angular ${angularType} named ${name}..`);
+    const out = await getRunOutput(`npx ionic generate ${angularType} ${name}`, project.projectFolder());
     const channel = getOutputChannel();
     channel.appendLine(out);
     const src = getStringFrom(out, 'CREATE ', '.ts');